requirements.txt
asyncio==3.4.3
mysql-connector-python==9.4.0
numpy==2.3.2
pandas==2.3.1
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.42
urllib3==2.5.0
isodate==0.7.2
py-clob-client==0.24.0
openai==1.99.9
beautifulsoup4==4.13.4
markdown==3.8.2

playwright>=1.40.0
dataclasses
typing
websocket-client==1.8.0
socketify==0.0.3
textdistance[extras]==4.6.3
requests[socks]
streamlit>=1.28.0
plotly>=5.17.0
